tiktok_auto/requirements.txt
playwright>=1.44.0
moviepy>=1.0.3
Pillow>=10.0.0
schedule>=1.2.0
numpy>=1.24.0
